refactored adapters layer
