+++ src/main/java/UI/ExitOptionsScreen.java
@@ -1,6 +1,6 @@
 package UI;
 
-import controller.StorylineController;
+import adapters.story.StorylineController;
 
 import javax.swing.*;
 import java.io.IOException;
