+++ src/main/java/adapters/gamesave/LoadInteractor.java
@@ -1,5 +1,6 @@
-package usecases.gamesave;
+package adapters.gamesave;
 import entities.player.PlayerData;
+import usecases.gamesave.Loader;
 
 import java.io.*;
