+++ src/main/java/usecases/login/LoginInteractor.java
@@ -1,7 +1,5 @@
 package usecases.login;
 
-import presenter.LoginInterface;
-
 import java.io.*;
 import java.util.HashMap;
 import java.util.Objects;
